analyzer/requirements.txt
anthropic>=0.42,<1.0
psycopg2-binary>=2.9,<3.0
rich>=13.0,<14.0
